requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1   
WTForms==3.1.2         
email-validator==2.2.0 
pyyaml==6.0.1   
markdown==3.7   
